Moves all components related to worker entity to a separate components folder. Modifies WorkerInventoryComponent.java to instead use code defined in UnitInventory.java.
