+++ source/core/src/main/com/deco2800/game/input/KeyboardInputFactory.java
@@ -1,7 +1,7 @@
 package com.deco2800.game.input;
 import com.deco2800.game.components.player.KeyboardPlayerInputComponent;
 import com.deco2800.game.ui.terminal.KeyboardTerminalInputComponent;
-import com.deco2800.game.worker.KeyboardWorkerInputComponent;
+import com.deco2800.game.worker.components.KeyboardWorkerInputComponent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
