+++ source/core/src/main/com/deco2800/game/input/TouchInputFactory.java
@@ -2,8 +2,7 @@ package com.deco2800.game.input;
 
 import com.deco2800.game.components.player.TouchPlayerInputComponent;
 import com.deco2800.game.ui.terminal.TouchTerminalInputComponent;
-import com.deco2800.game.worker.KeyboardWorkerInputComponent;
-import com.deco2800.game.worker.TouchWorkerInputComponent;
+import com.deco2800.game.worker.components.TouchWorkerInputComponent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
