+++ source/core/src/main/com/deco2800/game/worker/components/KeyboardWorkerInputComponent.java
@@ -1,19 +1,19 @@
-package com.deco2800.game.worker;
+package com.deco2800.game.worker.components;
 
-import com.badlogic.gdx.Input;
+import com.badlogic.gdx.Input.Keys;
 import com.badlogic.gdx.InputProcessor;
 import com.badlogic.gdx.math.Vector2;
 import com.deco2800.game.input.InputComponent;
 import com.deco2800.game.utils.math.Vector2Utils;
 
 /**
- * Input handler for the player for keyboard and touch (mouse) input.
- * This input handler uses keyboard and touch input.
+ * Input handler for the worker for keyboard and touch (mouse) input.
+ * This input handler only uses keyboard input.
  */
-public class TouchWorkerInputComponent extends InputComponent {
+public class KeyboardWorkerInputComponent extends InputComponent {
     private final Vector2 walkDirection = Vector2.Zero.cpy();
 
-    public TouchWorkerInputComponent() {
+    public KeyboardWorkerInputComponent() {
         super(5);
     }
 
@@ -26,21 +26,24 @@ public class TouchWorkerInputComponent extends InputComponent {
     @Override
     public boolean keyDown(int keycode) {
         switch (keycode) {
-            case Input.Keys.UP:
+            case Keys.W:
                 walkDirection.add(Vector2Utils.UP);
                 triggerWalkEvent();
                 return true;
-            case Input.Keys.LEFT:
+            case Keys.A:
                 walkDirection.add(Vector2Utils.LEFT);
                 triggerWalkEvent();
                 return true;
-            case Input.Keys.DOWN:
+            case Keys.S:
                 walkDirection.add(Vector2Utils.DOWN);
                 triggerWalkEvent();
                 return true;
-            case Input.Keys.RIGHT:
+            case Keys.D:
                 walkDirection.add(Vector2Utils.RIGHT);
                 triggerWalkEvent();
+                return true;
+            case Keys.SPACE:
+                entity.getEvents().trigger("attack");
                 return true;
             default:
                 return false;
@@ -56,19 +59,19 @@ public class TouchWorkerInputComponent extends InputComponent {
     @Override
     public boolean keyUp(int keycode) {
         switch (keycode) {
-            case Input.Keys.UP:
+            case Keys.W:
                 walkDirection.sub(Vector2Utils.UP);
                 triggerWalkEvent();
                 return true;
-            case Input.Keys.LEFT:
+            case Keys.A:
                 walkDirection.sub(Vector2Utils.LEFT);
                 triggerWalkEvent();
                 return true;
-            case Input.Keys.DOWN:
+            case Keys.S:
                 walkDirection.sub(Vector2Utils.DOWN);
                 triggerWalkEvent();
                 return true;
-            case Input.Keys.RIGHT:
+            case Keys.D:
                 walkDirection.sub(Vector2Utils.RIGHT);
                 triggerWalkEvent();
                 return true;
