+++ source/core/src/main/com/deco2800/game/worker/components/ResourceCollectComponent.java
@@ -1,4 +1,4 @@
-package com.deco2800.game.worker;
+package com.deco2800.game.worker.components;
 
 import com.badlogic.gdx.physics.box2d.Fixture;
 import com.deco2800.game.components.Component;
