+++ source/core/src/main/com/deco2800/game/worker/components/ResourceStatsComponent.java
@@ -1,4 +1,4 @@
-package com.deco2800.game.worker;
+package com.deco2800.game.worker.components;
 
 import com.deco2800.game.components.Component;
 import org.slf4j.Logger;
